refactor: extract GltfMaterialWriter, GltfMeshBuilder, GltfBufferUtils from JointedModelGltfExporter
Reduce JointedModelGltfExporter.java from 724 to 276 lines by extracting
three package-private delegates:

- GltfBufferUtils: normalize() and convertToFloatArray() statics
- GltfMaterialWriter: texture/material creation
- GltfMeshBuilder: mesh, skin, morph-target, and weight construction

Public API is unchanged; thin wrappers delegate to the new classes.
Adds GltfExporterCharacterizationTest (7 tests) pinning extracted behaviour.

Closes #611
